separete forgotPass.js in controller
